fix(emit): split ir printer config helpers
